requirements.txt
asteroid_filterbanks==0.4.0
auraloss==0.4.0
complextorch==1.2.0
einops==0.8.2
fairseq==0.12.2
jiwer==4.0.0
librosa==0.11.0
matplotlib==3.10.8
nltk==3.9.1
numpy==2.4.1
openai_whisper==20250625
pandas==3.0.0
pystoi==0.4.1
pytorch_lightning==2.5.0
PyYAML==6.0.2
PyYAML==6.0.3
Requests==2.32.5
scikit_learn==1.8.0
scipy==1.17.0
soundfile==0.13.1
torch==2.9.0
torchaudio==2.9.0
torchcrepe==0.0.23
tqdm==4.67.1
transformers==4.57.3
visqol==3.3.3
